refactor: creationDate를 uint32로 변경 (가스 최적화)
- ArtworkToken: creationDate 타입을 uint256에서 uint32로 변경하여 저장 비용 절감
- registerArtwork 시 파라미터 타입 업데이트
- 작품 제작 연도나 타임스탬프(초 단위)를 uint32로 저장하는 것이 적절함

+++ contracts/ArtworkToken.sol
@@ -10,7 +10,7 @@ contract ArtworkToken is ERC721, Ownable {
         string name;
         string artist;
         string location;
-        uint256 creationDate;
+        uint32 creationDate;
         uint256 pricePerPiece;
     }
 
@@ -24,7 +24,7 @@ contract ArtworkToken is ERC721, Ownable {
         string memory name,
         string memory artist,
         string memory location,
-        uint256 creationDate,
+        uint32 creationDate,
         uint256 pricePerPiece
     ) public onlyOwner returns (uint256) {
         require(bytes(name).length > 0, "Name is required");
